Add files via upload
Finished Filter class. Struggled a little bit with backend control flow. Still not sure how we are going to create the filters during runtime. Updated WeatherEvent class to have a type so we can pass that to Filter and we can process during runtime. Created FilterGenerator and FilterRepository for simple IO functionality for Filters as we have a filters table in the database. Time: 1hr 15 min